+++ pkg/server/grpc/option.go
@@ -7,6 +7,7 @@ import (
 
 type options struct {
 	gracePeriod time.Duration
+	maxConnAge  time.Duration
 	listen      string
 
 	tlsConfig *tls.Config
@@ -43,5 +44,12 @@ func WithListen(s string) Option {
 func WithTLSConfig(cfg *tls.Config) Option {
 	return optionFunc(func(o *options) {
 		o.tlsConfig = cfg
+	})
+}
+
+// WithMaxConnAge sets the maximum connection age for gRPC server.
+func WithMaxConnAge(t time.Duration) Option {
+	return optionFunc(func(o *options) {
+		o.maxConnAge = t
 	})
 }